# swagger.yaml
openapi: 3.0.3
info:
  title: Media server supports MPEG-DASH
  description: 
    DASH와 ClearKey 기반의 암호화를 지원하는 MPEG DASH 미디어 서버입니다.
  contact:
    email: dooheekim5@gmail.com
  version: 0.1.0
servers:
  - url: http://localhost:8080/
...
paths:
  /dashVideo/{filename}:
    get:
      tags:
        - Main
      summary: DASH를 이용하여 비디오 파일을 가져옵니다
      description: 재생을 위해 DASH를 이용하여 비디오 파일을 가져옵니다.
      operationId: playDashVideo
      parameters:
        - name: filename
          in: path
          description: DASH로 인코딩 된 영상을 가져옵니다. 파일은 ".mpd" 혹은 ".m4s" 확장자만 지원합니다.
          required: true
          schema:
            format: string.string
      responses:
        '200':
          description: Successful operation
          content:
            video/m4s:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid ID supplied
        '404':
          description: File not found
        '406':
          description: Given extension is not supported - Only supports mpd and m4s.
  /dashVideo:
    post:
      tags:
        - Main
      summary: 비디오를 업로드합니다
      description: 비디오를 업로드합니다
      operationId: uploadVideo
      requestBody:
        description: 비디오를 업로드합니다
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                exposure:
                  type: string
                  enum:
                    - PRIVATE
                    - PUBLIC
                  format: base64
                  example: PRIVATE
                title:
                  type: string
                  format: string
                  example: Big buck bunny
                description:
                  type: string
                  format: string
                  example: The plot follows a day in the life of Big Buck Bunny ... 
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '405':
          description: Invalid input
...
components:
  schemas:
    Video:
      required:
        - title
        - description
      type: object
      properties:
        id:
          type: string
          format: string
          example: random_video_id
        title:
          type: string
          format: string
          example: Big buck bunny
        description:
          type: string
          format: string
          example: The plot follows a day in the life of Big Buck Bunny...
...